requirements.txt
python-dotenv==1.0.1
vk-api==11.9.9